+++ stable_baselines/common/vec_env/__init__.py
@@ -6,3 +6,4 @@ from stable_baselines.common.vec_env.subproc_vec_env import SubprocVecEnv
 from stable_baselines.common.vec_env.vec_frame_stack import VecFrameStack
 from stable_baselines.common.vec_env.vec_normalize import VecNormalize
 from stable_baselines.common.vec_env.vec_video_recorder import VecVideoRecorder
+from stable_baselines.common.vec_env.vec_check_nan import VecCheckNan